input and inputfile
also some changes to combobox
