+++ app/src/main/java/reservant_mobile/ui/components/FormComponents.kt
@@ -1,0 +1,224 @@
+package reservant_mobile.ui.components
+
+import android.content.Context
+import android.net.Uri
+import androidx.activity.compose.rememberLauncherForActivityResult
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.interaction.PressInteraction
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.AttachFile
+import androidx.compose.material.icons.filled.Delete
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.focus.onFocusChanged
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontStyle
+import androidx.compose.ui.text.input.ImeAction
+import androidx.compose.ui.text.input.VisualTransformation
+import androidx.compose.ui.unit.dp
+import com.example.reservant_mobile.R
+import reservant_mobile.data.utils.getFileName
+
+@Composable
+fun FormInput(
+    modifier: Modifier = Modifier,
+    inputText: String,
+    onValueChange: (String) -> Unit,
+    label: String = "",
+    placeholder: String = "",
+    visualTransformation: VisualTransformation = VisualTransformation.None,
+    keyboardOptions: KeyboardOptions = KeyboardOptions.Default,
+    isError: Boolean = false,
+    errorText: String = "",
+    formSent: Boolean = false,
+    optional: Boolean = false,
+    maxLines: Int = 1,
+    leadingIcon: @Composable (() -> Unit)? = null,
+    shape: RoundedCornerShape = RoundedCornerShape(8.dp),
+) {
+
+    var beginValidation: Boolean by remember {
+        mutableStateOf(false)
+    }
+
+    if (inputText.isNotEmpty())
+        beginValidation = true
+
+    if (inputText.isEmpty() && optional)
+        beginValidation = false
+
+    Column {
+        OutlinedTextField(
+            modifier =
+            if (optional) {
+                modifier
+                    .fillMaxWidth()
+                    .padding(vertical = 8.dp)
+            } else {
+                modifier
+                    .fillMaxWidth()
+                    .padding(vertical = 8.dp)
+                    .onFocusChanged {
+                        if (it.hasFocus) beginValidation = true
+                    }
+            },
+            value = inputText,
+            onValueChange = onValueChange,
+            label = {
+                Row {
+                    Text(text = label)
+                    if (optional)
+                        Text(
+                            text = stringResource(id = R.string.label_optional),
+                            color = MaterialTheme.colorScheme.outline,
+                            fontStyle = FontStyle.Italic
+                        )
+
+                }
+            },
+            placeholder = { Text(text = placeholder) },
+            visualTransformation = visualTransformation,
+            keyboardOptions = keyboardOptions.copy(
+                imeAction = if (keyboardOptions.imeAction == ImeAction.Default)
+                    ImeAction.Next
+                else keyboardOptions.imeAction
+            ),
+            shape = shape,
+            isError = isError && (beginValidation || formSent),
+            maxLines = maxLines,
+            leadingIcon = leadingIcon,
+
+            )
+        if (isError && (beginValidation || formSent)) {
+            Text(
+                text = errorText,
+                color = MaterialTheme.colorScheme.error
+            )
+        }
+    }
+
+}
+
+@Composable
+fun FormFileInput(
+    label: String = "",
+    defaultValue: String = "",
+    onFilePicked: (Uri?) -> Unit,
+    modifier: Modifier = Modifier,
+    context: Context,
+    shape: RoundedCornerShape = RoundedCornerShape(8.dp),
+    isError: Boolean = false,
+    errorText: String = "",
+    formSent: Boolean = false,
+    optional: Boolean = false,
+    deletable: Boolean = false
+) {
+    var fileName by remember { mutableStateOf<String?>(null) }
+    val pickFileLauncher = rememberLauncherForActivityResult(
+        contract = ActivityResultContracts.GetContent()
+    ) { uri: Uri? ->
+        fileName = uri?.let { getFileName(context, it.toString()) }
+        onFilePicked(uri)
+    }
+    var beginValidation: Boolean by remember { mutableStateOf(false) }
+
+    if (fileName != null) {
+        beginValidation = true
+    }
+    if (fileName == null && optional) {
+        beginValidation = false
+    }
+
+    if (fileName == null && defaultValue.isNotBlank()){
+        fileName = defaultValue
+    }
+
+    OutlinedTextField(
+        modifier =
+        if (optional) {
+            modifier
+                .fillMaxWidth()
+                .padding(vertical = 8.dp)
+        } else {
+            modifier
+                .fillMaxWidth()
+                .padding(vertical = 8.dp)
+                .onFocusChanged {
+                    if (it.hasFocus) {
+                        beginValidation = true
+                    }
+                }
+        },
+        value = fileName ?: "",
+        onValueChange = { },
+        label = {
+            Row {
+                Text(text = label)
+                if (optional) {
+                    Text(
+                        text = stringResource(R.string.label_optional),
+                        color = Color.Gray,
+                        fontStyle = FontStyle.Italic
+                    )
+                }
+            }
+        },
+        readOnly = true,
+        visualTransformation = VisualTransformation.None,
+        keyboardOptions = KeyboardOptions.Default,
+        interactionSource = remember { MutableInteractionSource() }
+            .also { interactionSource ->
+                LaunchedEffect(interactionSource) {
+                    interactionSource.interactions.collect {
+                        if (it is PressInteraction.Release) {
+                            pickFileLauncher.launch("*/*")
+                        }
+                    }
+                }
+            },
+        trailingIcon = {
+            if (deletable && fileName != null) {
+                Icon(
+                    imageVector = Icons.Default.Delete,
+                    contentDescription = "Delete file",
+                    modifier = Modifier.clickable {
+                        fileName = null
+                        onFilePicked(null)
+                    }
+                )
+            } else if (fileName == null) {
+                Icon(
+                    imageVector = Icons.Default.AttachFile,
+                    contentDescription = "Attach file"
+                )
+            }
+        },
+        shape = shape,
+        isError = isError && (beginValidation || formSent),
+    )
+
+    if (isError && (beginValidation || formSent)) {
+        Text(
+            text = errorText,
+            color = Color.Red
+        )
+    }
+}
